Refactor: EmployeeStatus enums 패키지로 이동 [#30]
